Rename module to fix naming case problem
